+++ kilo/management/commands/import_workouts.py
@@ -5,7 +5,7 @@ import re
 
 from django.core.management.base import BaseCommand, CommandError
 
-from miles.models import Day Workout
+from kilo.models import Day Workout
 
 
 class Command(BaseCommand):
