+++ kilo/templates/kilo/days.html
@@ -1,6 +1,6 @@
 {% extends 'orange/base.html' %}
 
-{% block title %}Miles{% endblock title %}
+{% block title %}Kilo{% endblock title %}
 
 {% block style %}{{ block.super }}
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
@@ -21,7 +21,7 @@
           crossorigin="anonymous">
   </script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/knockout/3.5.0/knockout-min.js"></script>
-  <script src="/static/miles/js/miles.js"></script>
+  <script src="/static/kilo/js/kilo.js"></script>
 {% endblock js %}
 
 {% block body %}
