+++ orange/settings.py
@@ -34,7 +34,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'miles',
+    'kilo',
     'orange',
     'registration',
     'rhyme',
